+++ src/NadekoBot/Modules/Searches/Services/AnimeSearchService.cs
@@ -50,7 +50,6 @@ namespace NadekoBot.Modules.Searches.Services
             }
         }
 
-        // todo fix novel
         public async Task<NovelResult> GetNovelData(string query)
         {
             if (string.IsNullOrWhiteSpace(query))